+++ dataiter/dt.py
@@ -25,6 +25,7 @@ import numpy as np
 
 from dataiter import util
 from dataiter import Vector
+from numpy.dtypes import StringDType
 
 def day(x):
     """
@@ -46,11 +47,11 @@ def from_string(x, format):
     >>> dt.from_string(x, "%d.%m.%Y")
     """
     if util.is_scalar(x):
-        x = np.array([x], str)
+        x = Vector([x], str)
         return from_string(x, format)[0]
     x = util.sequencify(x)
     assert isinstance(x, np.ndarray)
-    assert np.issubdtype(x.dtype, np.str_)
+    assert isinstance(x.dtype, StringDType)
     out = np.full_like(x, None, object)
     out = Vector.fast(out, object)
     na = x == ""
@@ -146,7 +147,7 @@ def now():
 
 def _pull_datetime(x, function):
     if util.is_scalar(x):
-        x = np.array([x], np.datetime64)
+        x = Vector([x], np.datetime64)
         return _pull_datetime(x, function)[0]
     x = util.sequencify(x)
     assert isinstance(x, np.ndarray)
@@ -161,7 +162,7 @@ def _pull_datetime(x, function):
 
 def _pull_int(x, function):
     if util.is_scalar(x):
-        x = np.array([x], np.datetime64)
+        x = Vector([x], np.datetime64)
         return _pull_int(x, function)[0]
     x = util.sequencify(x)
     assert isinstance(x, np.ndarray)
@@ -176,7 +177,7 @@ def _pull_int(x, function):
 
 def _pull_str(x, function):
     if util.is_scalar(x):
-        x = np.array([x], np.datetime64)
+        x = Vector([x], np.datetime64)
         return _pull_str(x, function)[0]
     x = util.sequencify(x)
     assert isinstance(x, np.ndarray)