+++ dataiter/list_of_dicts.py
@@ -33,7 +33,6 @@ import sys
 from attd import AttributeDict
 from dataiter import deco
 from dataiter import util
-from math import inf
 
 class ListOfDicts(list):
 
@@ -782,27 +781,15 @@ class ListOfDicts(list):
     def _to_columns(self):
         return {k: self.pluck(k) for k in self[0]} if self else {}
 
-    def to_data_frame(self, strings_as_object=inf):
+    def to_data_frame(self):
         """
         Return list converted to a :class:`.DataFrame`.
-
-        `strings_as_object` is a cutoff point. If any row has more characters
-        than that, the whole column will use the object data type. This is
-        intended to help limit memory use as NumPy strings are fixed-length and
-        can take a huge amount of memory if even a single row is long.
 
         >>> data = di.read_json("data/listings.json")
         >>> data.to_data_frame()
         """
         from dataiter import DataFrame
-        from dataiter import DataFrameColumn
         data = self._to_columns()
-        if strings_as_object < inf:
-            for name in data:
-                if (data[name] and
-                    any(isinstance(x, str) for x in data[name]) and
-                    any(len(x) > strings_as_object for x in data[name] if isinstance(x, str))):
-                    data[name] = DataFrameColumn(data[name], object)
         return DataFrame(**data)
 
     def to_json(self, **kwargs):